requirements-dev.txt
google-generativeai
supabase
yt-dlp
youtube-transcript-api
beautifulsoup4
aiohttp
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
tiktoken
rank_bm25
